requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.30.0
better-bing-image-downloader>=3.0.0
Pillow>=10.0.0
google-genai>=1.0.0
